fellowboard, fellowmatching 폴더 fellow로 합침. fellow-mapper 추가, mybatis-config.xml 수정
