+++ src/main/java/com/a2b/trip/fellow/controller/FellowBoardController.java
@@ -1,9 +1,9 @@
-package com.a2b.trip.fellowboard.controller;
+package com.a2b.trip.fellow.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 
-import com.a2b.trip.fellowboard.model.service.FellowBoardService;
+import com.a2b.trip.fellow.model.service.FellowBoardService;
 
 @Controller
 public class FellowBoardController {
