+++ src/main/java/com/a2b/trip/fellow/controller/FellowMatchingController.java
@@ -1,9 +1,9 @@
-package com.a2b.trip.fellowmatching.controller;
+package com.a2b.trip.fellow.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 
-import com.a2b.trip.fellowmatching.model.service.FellowMatchingService;
+import com.a2b.trip.fellow.model.service.FellowMatchingService;
 
 @Controller
 public class FellowMatchingController {
