+++ src/main/java/com/a2b/trip/fellow/model/dao/FellowBoardDao.java
@@ -1,14 +1,14 @@
-package com.a2b.trip.fellowmatching.model.dao;
+package com.a2b.trip.fellow.model.dao;
 
 import org.mybatis.spring.SqlSessionTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
-@Repository("fellowMatchingDao")
-public class FellowMatchingDao {
+@Repository("fellowBoardDao")
+public class FellowBoardDao {
 
 	@Autowired
 	private SqlSessionTemplate sqlSession;
 	
-	public FellowMatchingDao() {}
+	public FellowBoardDao() {}
 }
