+++ src/main/java/com/a2b/trip/fellow/model/dao/FellowMatchingDao.java
@@ -1,14 +1,14 @@
-package com.a2b.trip.fellowboard.model.dao;
+package com.a2b.trip.fellow.model.dao;
 
 import org.mybatis.spring.SqlSessionTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
-@Repository("fellowBoardDao")
-public class FellowBoardDao {
+@Repository("fellowMatchingDao")
+public class FellowMatchingDao {
 
 	@Autowired
 	private SqlSessionTemplate sqlSession;
 	
-	public FellowBoardDao() {}
+	public FellowMatchingDao() {}
 }
