+++ src/main/java/com/a2b/trip/fellow/model/service/FellowBoardService.java
@@ -1,0 +1,5 @@
+package com.a2b.trip.fellow.model.service;
+
+public interface FellowBoardService {
+
+}
