+++ src/main/java/com/a2b/trip/fellow/model/service/FellowMatchingService.java
@@ -1,0 +1,5 @@
+package com.a2b.trip.fellow.model.service;
+
+public interface FellowMatchingService {
+
+}
