+++ src/main/java/com/a2b/trip/fellow/model/vo/FellowMatching.java
@@ -1,4 +1,4 @@
-package com.a2b.trip.fellowmatching.model.vo;
+package com.a2b.trip.fellow.model.vo;
 
 import java.io.Serializable;
 import java.sql.Date;
