--- src/main/java/com/a2b/trip/fellowboard/model/service/FellowBoardService.java
@@ -1,5 +1,0 @@
-package com.a2b.trip.fellowboard.model.service;
-
-public interface FellowBoardService {
-
-}
