--- src/main/java/com/a2b/trip/fellowmatching/model/service/FellowMatchingService.java
@@ -1,5 +1,0 @@
-package com.a2b.trip.fellowmatching.model.service;
-
-public interface FellowMatchingService {
-
-}
