+++ src/main/resources/mappers/fellow-mapper.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!DOCTYPE mapper PUBLIC
+"-//mybatis.org//DTD Mapper 3.0//EN"
+"http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="fellowMapper">
+	
+	<select id="selectMyFellowMatching" parameterType="string" resultType="FellowMatching">
+		select * from fellowmatching where fm_id = #{value}
+	</select>
+	
+</mapper>
