+++ src/main/resources/mybatis-config.xml
@@ -14,6 +14,7 @@
 		<typeAlias type="com.a2b.trip.guide.model.vo.Guide" alias="Guide"/>
 		<typeAlias type="com.a2b.trip.guide.model.vo.GuideDetail" alias="GuideDetail"/>
 		<typeAlias type="com.a2b.trip.notice.model.vo.Notice" alias="Notice"/>
+		<typeAlias type="com.a2b.trip.notice.model.vo.FellowMatching" alias="FellowMatching" />
 	</typeAliases>
 	
 	<mappers>
@@ -22,6 +23,7 @@
 		<mapper resource="mappers/qna-mapper.xml" />
 		<mapper resource="mappers/guide-mapper.xml" />
 		<mapper resource="mappers/notice-mapper.xml"/>
+		<mapper resource="mappers/fellow-mapper.xml" />
 		
 	</mappers>
 </configuration>
